Refactored the build_ngtf.py by creating tools/build_utils.py and updating build_ngtf.py to refer to the former. Also added an option to specific the artifacts directory (ourside of the build tree)
