+++ test_ngtf.py
@@ -24,7 +24,7 @@ import glob
 import platform
 from distutils.sysconfig import get_python_lib
 
-from build_ngtf import load_venv, command_executor
+from tools.build_utils import load_venv, command_executor
 
 
 def run_ngtf_gtests(build_dir, filters):
