--- tools/CMakeLists.txt
@@ -1,38 +1,0 @@
-# Copyright 2018 Nervana Systems Inc.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-set(LIB_NAME graph_analyzer)
-
-include_directories(${TensorFlow_INCLUDE_DIR})
-include_directories(${TensorFlow_INCLUDE_DIR}/external/nsync/public)
-
-# For some reason the following is needed for ABI compatibility with TF.
-# (There must be some dependency on it in the struct/class definitions.)
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DNDEBUG")
-
-#-----------------------------------------------------------------------------------------------
-# Compiler-specific logic...
-#-----------------------------------------------------------------------------------------------
-set(SRC 
-   tf_graph_analyzer.cc
-)
-
-add_library(${LIB_NAME} SHARED ${SRC})
-target_link_libraries( ${LIB_NAME} tensorflow_framework_lib )
-target_link_libraries( ${LIB_NAME} ngraph_logger) 
-
-target_compile_definitions( 
-    ${LIB_NAME} PRIVATE
-    _GLIBCXX_USE_CXX11_ABI=${TensorFlow_CXX_ABI}
-)
-set_property(TARGET ${LIB_NAME} PROPERTY POSITION_INDEPENDENT_CODE ON)
